Last night's late night bug fixes
When the Redis server goes away, the redis client was caching all
requests.  This sounds cool, but it causes get/set callbacks to
not fire until the server itself picks back up and responds.

This means that, rather than reverting to a proxy, we were just holding
connection open to each client.  This is not ideal.  So I fixed it.